endpoint: move `NewEndpointWithState` to `test_utils.go`
Signed-off by: Ian Vernon <ian@cilium.io>
