+++ pkg/endpoint/endpoint.go
@@ -263,38 +263,6 @@ func (e *Endpoint) LXCMac() mac.MAC {
 	return e.mac
 }
 
-// NewEndpointWithState creates a new endpoint useful for testing purposes
-func NewEndpointWithState(owner regeneration.Owner, proxy EndpointProxy, allocator cache.IdentityAllocator, ID uint16, state string) *Endpoint {
-	ep := &Endpoint{
-		owner:           owner,
-		proxy:           proxy,
-		ID:              ID,
-		OpLabels:        pkgLabels.NewOpLabels(),
-		status:          NewEndpointStatus(),
-		DNSHistory:      fqdn.NewDNSCacheWithLimit(option.Config.ToFQDNsMinTTL, option.Config.ToFQDNsMaxIPsPerHost),
-		state:           state,
-		hasBPFProgram:   make(chan struct{}, 0),
-		controllers:     controller.NewManager(),
-		eventQueue:      eventqueue.NewEventQueueBuffered(fmt.Sprintf("endpoint-%d", ID), option.Config.EndpointQueueSize),
-		desiredPolicy:   policy.NewEndpointPolicy(owner.GetPolicyRepository()),
-		regenFailedChan: make(chan struct{}, 1),
-		allocator:       allocator,
-	}
-
-	ctx, cancel := context.WithCancel(context.Background())
-	ep.aliveCancel = cancel
-	ep.aliveCtx = ctx
-	ep.startRegenerationFailureHandler()
-	ep.realizedPolicy = ep.desiredPolicy
-
-	ep.SetDefaultOpts(option.Config.Opts)
-	ep.UpdateLogger(nil)
-
-	ep.eventQueue.Run()
-
-	return ep
-}
-
 // GetID returns the endpoint's ID as a 64-bit unsigned integer.
 func (e *Endpoint) GetID() uint64 {
 	return uint64(e.ID)
